python/requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
docker==2.4.2
docker-pycreds==0.2.1
idna==2.5
requests==2.18.2
six==1.10.0
urllib3==1.22
websocket-client==0.44.0
contexttimer==0.3.3
PyYAML==3.12
Jinja2==2.10
cookiecutter==1.6.0
